Phase 1: optimized entity design & code fixes
