--- slotbooker/src/main/java/com/deepan/slotbooker/model/BookingStatus.java
@@ -1,6 +1,0 @@
-package com.deepan.slotbooker.model;
-
-public enum BookingStatus {
-    BOOKED,
-    CANCELLED
-}
