+++ slotbooker/src/main/java/com/deepan/slotbooker/model/Facility.java
@@ -6,6 +6,10 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import org.hibernate.annotations.CreationTimestamp;
+import org.hibernate.annotations.UpdateTimestamp;
+
+import java.time.LocalDateTime;
 
 @Entity
 @Table (name = "facility")
@@ -32,5 +36,11 @@ public class Facility {
     private Sport sport;
 
     private Double pricePerHour;
+
+    @CreationTimestamp
+    private LocalDateTime createdTime;
+
+    @UpdateTimestamp
+    private LocalDateTime updatedTime;
 
 }
