--- slotbooker/src/main/java/com/deepan/slotbooker/model/Role.java
@@ -1,6 +1,0 @@
-package com.deepan.slotbooker.model;
-
-public enum Role {
-    PLAYER,
-    OWNER
-}
