+++ slotbooker/src/main/java/com/deepan/slotbooker/model/User.java
@@ -1,10 +1,18 @@
 package com.deepan.slotbooker.model;
 
+import com.deepan.slotbooker.model.enums.Role;
+import com.deepan.slotbooker.model.enums.UserStatus;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import org.hibernate.annotations.CreationTimestamp;
+import org.hibernate.annotations.UpdateTimestamp;
+import org.springframework.data.annotation.CreatedDate;
+
+import java.time.LocalDateTime;
+import java.util.Date;
 
 @Entity
 @Table(name = "users")
@@ -31,5 +39,13 @@ public class User {
     @Enumerated(EnumType.STRING)
     private Role userRole;
 
+    @Enumerated(EnumType.STRING)
+    private UserStatus userStatus;
+
+    @CreationTimestamp
+    private LocalDateTime createdTime;
+
+    @UpdateTimestamp
+    private LocalDateTime updatedTime;
 
 }
