+++ slotbooker/src/main/java/com/deepan/slotbooker/model/Venue.java
@@ -3,6 +3,10 @@ package com.deepan.slotbooker.model;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import jakarta.persistence.*;
 import lombok.*;
+import org.hibernate.annotations.CreationTimestamp;
+import org.hibernate.annotations.UpdateTimestamp;
+
+import java.time.LocalDateTime;
 
 @Entity
 @Table (name = "venue")
@@ -20,8 +24,18 @@ public class Venue {
 
     private String location;
 
+    private String description;
+
+    private String contactInfo;
+
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "owner_id", nullable = false)
     @JsonIgnoreProperties({"hibernateLazyInitializer", "handler"})
     private User ownerId;
+
+    @CreationTimestamp
+    private LocalDateTime createdTime;
+
+    @UpdateTimestamp
+    private LocalDateTime updatedTime;
 }
