+++ slotbooker/src/main/java/com/deepan/slotbooker/model/enums/BookingStatus.java
@@ -1,0 +1,7 @@
+package com.deepan.slotbooker.model.enums;
+
+public enum BookingStatus {
+    BOOKED,
+    CANCELLED,
+    RESERVED
+}
