+++ slotbooker/src/main/java/com/deepan/slotbooker/model/enums/Role.java
@@ -1,0 +1,7 @@
+package com.deepan.slotbooker.model.enums;
+
+public enum Role {
+    PLAYER,
+    OWNER,
+    ADMIN
+}
