+++ slotbooker/src/main/java/com/deepan/slotbooker/model/enums/UserStatus.java
@@ -1,0 +1,5 @@
+package com.deepan.slotbooker.model.enums;
+
+public enum UserStatus {
+    ACTIVE, INACTIVE, SUSPENDED
+}
