+++ slotbooker/src/main/java/com/deepan/slotbooker/security/UserSecurity.java
@@ -26,7 +26,7 @@ public class UserSecurity implements UserDetails {
 
     @Override
     public String getUsername() {
-        return user.getUserName();
+        return user.getMobileNumber();
     }
 
     @Override
