+++ slotbooker/src/test/java/com/deepan/slotbooker/service/BookingServiceTest.java
@@ -5,7 +5,7 @@ import com.deepan.slotbooker.dto.booking.BookingResponse;
 import com.deepan.slotbooker.exception.ResourceNotFoundException;
 import com.deepan.slotbooker.mapper.BookingMapper;
 import com.deepan.slotbooker.model.Booking;
-import com.deepan.slotbooker.model.BookingStatus;
+import com.deepan.slotbooker.model.enums.BookingStatus;
 import com.deepan.slotbooker.model.Slot;
 import com.deepan.slotbooker.model.User;
 import com.deepan.slotbooker.repository.BookingRepository;
